Create utilities for google servers
